[review] split quantizer tests by module; clarify copy=True
#A: residual_quantizer_test.py tested both the base ResidualQuantizer and
    ResidualKMeansQuantizer. Split the K-Means tests into the matching
    residual_kmeans_quantizer_test.py (so each module has its own test file);
    the base tests stay put.
#B: expand the train_offline copy=True comment — the residual loop mutates x
    in place and the input is a view into the reservoir buffer, so it must own
    a fresh copy (copy=True is a single guaranteed copy vs a double-copy clone).
